requirements.txt
appdirs==1.4.3
argon2-cffi==19.2.0
attrs==19.3.0
black==19.10b0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
dnspython==1.16.0
Flask==1.1.1
Flask-Cors==3.0.8
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mypy==0.761
mypy-extensions==0.4.3
pathspec==0.7.0
pycparser==2.19
pymongo==3.9.0
regex==2020.2.20
requests==2.22.0
six==1.13.0
toml==0.10.0
typed-ast==1.4.1
typing-extensions==3.7.4.1
urllib3==1.25.7
Werkzeug==0.16.0
yapf==0.29.0